fix: match DTO constructor to JPQL result types (quantity as Long)

+++ src/main/java/com/pl/example/order/dto/BookOrderReportItem.java
@@ -2,5 +2,5 @@ package com.pl.example.order.dto;
 
 public record BookOrderReportItem(String isbn,
                                   String title,
-                                  Integer quantity) {
+                                  Long quantity) {
 }
